libgps updates
git-svn-id: https://svn.code.sf.net/p/drm/code@1349 ddb839e9-6489-4ed3-a1ee-f9006aeb25b6

+++ branches/dream-ollie/src/util/Hamlib.cpp
@@ -424,7 +424,7 @@ CHamlib::LoadSettings(CSettings & s)
 		SetHamlibModelID(model);
 	}
 
-	s.Put("Hamlib", "hamlib-model", model);
+    s.Put("Hamlib", "hamlib-model", (int)model);
 	s.Put("Hamlib", "hamlib-config", strHamlibConf);
 	s.Put("Hamlib", "settings", strSettings);
 	s.Put("Hamlib", "freqoffset", iFreqOffset);
@@ -434,7 +434,7 @@ void
 CHamlib::SaveSettings(CSettings & s)
 {
 	/* Hamlib Model ID */
-	s.Put("Hamlib", "hamlib-model", iHamlibModelID);
+    s.Put("Hamlib", "hamlib-model", (int)iHamlibModelID);
 
 	/* Hamlib configuration string */
 	stringstream ss;
